entrance-system/musaade/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
cssselect2==0.4.1
Django==4.0
django-htmlmin==0.11.0
djangorestframework==3.13.1
fonttools==4.28.5
gunicorn==20.1.0
html5lib==1.1
humanfriendly==10.0
idna==3.3
Naked==0.1.31
Pillow==9.0.0
psycopg2-binary==2.9.2
pycparser==2.21
pycryptodome==3.12.0
pydyf==0.1.2
pyphen==0.12.0
pypng==0.0.21
PyQRCode==1.2.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
shellescape==3.8.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tinycss2==1.1.1
urllib3==1.26.7
weasyprint==53.4
webencodings==0.5.1
xlwt==1.3.0
zopfli==0.1.9
